# xswagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LAARTESA API",
    "version": "1.0.0",
    "description": "Documentación de la API para WEB APP LA ARTESA"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor Local"
    },
    {
      "url": "https://f686-160-173-184-21.ngrok-free.app",
      "description": "Servidor Ngrok"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ingrese el token JWT en el formato: Bearer <token>"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priceList1": {
            "type": "number"
          },
          "priceList2": {
            "type": "number"
          },
          "priceList3": {
            "type": "number"
          },
          "stock": {
            "type": "integer"
          },
          "barcode": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateProductRequest": {
        "type": "object",
        "required": ["name", "priceList1", "stock"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Producto Ejemplo"
          },
          "description": {
            "type": "string",
            "example": "Descripción detallada del producto"
          },
          "priceList1": {
            "type": "number",
            "example": 100.00
          },
          "priceList2": {
            "type": "number",
            "example": 90.00
          },
          "priceList3": {
            "type": "number",
            "example": 80.00
          },
          "stock": {
            "type": "integer",
            "example": 50
          },
          "barcode": {
            "type": "string",
            "example": "7501234567890"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/images/product.jpg"
          }
        }
      },
      "UpdateProductImageRequest": {
        "type": "object",
        "required": ["imageUrl"],
        "properties": {
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/images/updated-product.jpg"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Token de acceso no proporcionado o inválido",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Se requiere un token de acceso"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "No tiene permisos suficientes para acceder al recurso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "No tiene permisos para realizar esta acción"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/products": {
      "post": {
        "tags": ["Products"],
        "summary": "Crear un nuevo producto",
        "description": "Crea un nuevo producto en el sistema. Requiere autenticación y rol de administrador.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "get": {
        "tags": ["Products"],
        "summary": "Obtener todos los productos",
        "description": "Retorna la lista de todos los productos. Requiere autenticación.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de productos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/products/{productId}": {
      "get": {
        "tags": ["Products"],
        "summary": "Obtener un producto por ID",
        "description": "Retorna los detalles de un producto específico. Requiere autenticación.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "description": "ID del producto a consultar",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del producto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/products/{productId}/image": {
      "put": {
        "tags": ["Products"],
        "summary": "Actualizar la imagen de un producto",
        "description": "Actualiza la URL de la imagen de un producto. Requiere autenticación y rol de administrador.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "description": "ID del producto a actualizar",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductImageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagen del producto actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
  }
}